Moved literal, verifyprimitives to primitives.js
